Set collector w/test
