+++ src/test/java/bonus/SetCollectorTest.java
@@ -1,31 +1,21 @@
 package bonus;
 
-import org.openjdk.jmh.annotations.*;
+import org.junit.Test;
 
-import java.util.Arrays;
 import java.util.Set;
-import java.util.concurrent.ThreadLocalRandom;
-import java.util.concurrent.TimeUnit;
-import java.util.stream.IntStream;
 
-@Fork(5)
-@BenchmarkMode(Mode.AverageTime)
-@OutputTimeUnit(TimeUnit.MILLISECONDS)
-//@Warmup(iterations = 20, time = 1, timeUnit = TimeUnit.MINUTES)
-//@Measurement(iterations = 20, time = 1, timeUnit = TimeUnit.MINUTES)
-@State(Scope.Thread)
+import static org.hamcrest.Matchers.is;
+import static org.junit.Assert.assertThat;
+
 public class SetCollectorTest {
-    private Integer[] millionInts = IntStream.generate(() -> ThreadLocalRandom.current().nextInt(1_000_000)).limit(1_000_000)
-            .boxed().toArray(Integer[]::new);
+    SetCollectorBenchmark scb = new SetCollectorBenchmark();
 
-    @Benchmark
-    public Set<Integer> collectSet() {
-        return Arrays.stream(millionInts).collect(new CollectorToSet<>());
+    @Test
+    public void testSetCollection() {
+        final Set<Integer> set1 = scb.collectSet();
+        final Set<Integer> set2 = scb.collectConcurrentSet();
+        assertThat(set1, is(set2));
     }
 
-    @Benchmark
-    public Set<Integer> collectConcurrentSet() {
-        return Arrays.stream(millionInts).collect(new ConcurrentCollectorToSet<>());
-    }
 
 }
